# api/openapi.yaml
openapi: 3.1.0
info:
  title: PALISADE HTTP API
  version: 0.1.0
  description: Early behavior-first decision API. The runtime defaults to shadow mode; production decision and event calls use short-lived, session/action-bound proofs.
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
servers:
  - url: /
paths:
...
  /v1/session:
    post:
      summary: Issue a server-bound continuity cookie
      description: Backend-only endpoint. The integrating origin forwards the Set-Cookie header to the browser; cookie integrity establishes continuity, not humanity or account identity.
      operationId: issueSession
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Signed session cookie issued.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/token:
    post:
      summary: Issue a short-lived one-time proof
      operationId: issueProof
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [action]
              properties:
                session_id:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: Optional when a valid PALISADE session cookie is supplied; if both are present they must match.
                action: { type: string, minLength: 1, maxLength: 80 }
                ttl_seconds: { type: integer, minimum: 1, maximum: 300, default: 60 }
      responses:
        "201":
          description: One-time proof issued.
          content:
            application/json:
              schema:
                type: object
                required: [proof_token, expires_in]
                properties:
                  proof_token: { type: string }
                  expires_in: { type: integer }
        "401": { $ref: "#/components/responses/Error" }
...
  /v1/origin-check:
    post:
      summary: Evaluate and return the bounded origin enforcement result
      description: Alternative to /v1/decision for an origin or reverse-proxy subrequest. Production requires the one-time proof_token in the request body; signed rollouts additionally require the session cookie. A valid signed rollout may return 429 for throttle or 403 for challenge/block. Shadow and excluded canary requests return 204. Scores and evidence are not reflected in this response.
      operationId: checkOrigin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecisionRequest" }
      responses:
        "204":
          description: Pass the request. This includes shadow mode and requests excluded from a canary.
          headers:
            X-Palisade-Decision-ID: { schema: { type: string } }
            X-Palisade-Action: { schema: { type: string, enum: [allow, observe] } }
            X-Palisade-Handling: { schema: { type: string, enum: [pass] } }
            X-Palisade-Mode: { schema: { type: string, enum: [shadow, canary, enforce] } }
            X-Palisade-Rollout-ID: { schema: { type: string } }
        "403":
          description: Apply the challenge or temporary block selected by a valid signed rollout.
          headers:
            X-Palisade-Decision-ID: { schema: { type: string } }
            X-Palisade-Action: { schema: { type: string, enum: [challenge, block] } }
            X-Palisade-Handling: { schema: { type: string, enum: [challenge, block] } }
            X-Palisade-Mode: { schema: { type: string, enum: [canary, enforce] } }
            X-Palisade-Rollout-ID: { schema: { type: string } }
            X-Palisade-Challenge-ID:
              description: Present only for challenge handling. Retrieve the bound challenge at the Location route.
              schema: { type: string, minLength: 32, maxLength: 32 }
            Location:
              description: Present only for challenge handling.
              schema: { type: string, pattern: "^/v1/challenge/[A-Za-z0-9_-]{32}$" }
            Retry-After: { schema: { type: integer, minimum: 1, maximum: 3600 } }
        "429":
          description: Apply a one-second delay response or a bounded throttle selected by a valid signed rollout. PALISADE returns immediately and never sleeps in the hot path.
          headers:
            X-Palisade-Decision-ID: { schema: { type: string } }
            X-Palisade-Action: { schema: { type: string, enum: [delay, throttle] } }
            X-Palisade-Handling: { schema: { type: string, enum: [delay, throttle] } }
            X-Palisade-Mode: { schema: { type: string, enum: [canary, enforce] } }
            X-Palisade-Rollout-ID: { schema: { type: string } }
            Retry-After: { schema: { type: integer, minimum: 1, maximum: 60 } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminBearer:
      type: http
      scheme: bearer
      description: Distinct PALISADE_ADMIN_KEY accepted only by the loopback administrative listener.
    eventProof:
      type: apiKey
      in: header
      name: X-Palisade-Proof
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-palisade_session
  responses:
    Error:
      description: Stable machine-readable error.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
  schemas:
...
    SessionResponse:
      type: object
      additionalProperties: false
      required: [session_id, expires_at]
      properties:
        session_id: { type: string, minLength: 32, maxLength: 32 }
        expires_at: { type: string, format: date-time }
...
    Observations:
      type: object
      additionalProperties: false
      properties:
        user_agent_present: { type: boolean }
        browser_event_count:
          type: integer
          minimum: 0
          maximum: 10000
          description: Claimed count for compatibility. The live service replaces it with its bounded server-side event-store count before evaluation; an unverified claim never creates benign continuity evidence.
        honeypot_hits: { type: integer, minimum: 0, maximum: 100 }
        challenge_verdict:
          type: string
          enum: [suspicious, failed, blocked, allowed, passed, unknown]
          description: Normalized result from a trusted challenge adapter. A passed challenge is an outcome, not proof that the client is human.
        external_risk_score: { type: number, minimum: 0, maximum: 1 }
        policy_alert:
          type: boolean
          description: Trusted deployment policy alert normalized to a boolean; clients must not set this directly.
        verified_bot: { type: boolean }
        transport_protocol:
          type: string
          enum: [http1, http2, http3, unknown]
          description: Closed protocol family normalized by the trusted origin adapter; never a raw protocol string.
        transport_security:
          type: string
          enum: [direct_tls, trusted_proxy_tls, plaintext, unknown]
          description: Closed transport-security class. Forwarded TLS metadata is accepted only from a configured trusted TCP peer.
        client_address_source:
          type: string
          enum: [direct, trusted_proxy, invalid_trusted_proxy, unknown]
          description: Provenance class only. PALISADE does not receive the client or proxy address.
    DecisionRequest:
      type: object
      additionalProperties: false
      required: [action, endpoint_class, sequence, observations]
      properties:
        session_id:
          type: string
          minLength: 8
          maxLength: 128
          description: Optional when a valid PALISADE session cookie is supplied; if both are present they must match.
        action: { type: string, minLength: 1, maxLength: 80 }
        endpoint_class:
          type: string
          enum: [public_content, compare_index, compare_noindex, challenge_worker, other_public, account, login, checkout, other]
        sequence: { type: integer, minimum: 1 }
        proof_token: { type: string }
        observations: { $ref: "#/components/schemas/Observations" }
        evaluation_cohort:
          type: string
          enum: [standard, reduced_motion, keyboard_only, fallback_path, sensor_missing, unknown]
          description: Optional trusted, coarse operational cohort used only for aggregate evaluation. PALISADE does not infer identity, disability or demographic attributes; omitted values become unknown.
...